Testing prompt - working with compelter in scope of main file
